Added own regions, cleaned up code
Because of issues with interpolation, both GFED and box regions are available now. Box regions should eliminate the issue of defining regions in different resolutions. Renamed Inferno to INFERNO for consistency, and removed all mapping functions from model scripts as there is a general mapping script in spatial_comparison.py.